chore: renamed package ui to view
