+++ app/src/main/AndroidManifest.xml
@@ -14,7 +14,7 @@
         android:usesCleartextTraffic="true"
         tools:targetApi="m">
         <activity
-            android:name="com.example.vinilosapp.ui.MainActivity"
+            android:name="com.example.vinilosapp.view.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.VinilosApp.NoActionBar">
