+++ app/src/main/res/layout/album_fragment.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.AlbumFragment"
+    tools:context=".view.AlbumFragment"
     android:layout_marginStart="16dp"
     android:layout_marginEnd="16dp">
