+++ app/src/main/res/layout/comment_fragment.xml
@@ -6,7 +6,7 @@
     android:layout_marginStart="16dp"
     android:layout_marginEnd="16dp"
     android:orientation="vertical"
-    tools:context=".ui.CommentFragment">
+    tools:context=".view.CommentFragment">
 
     <TextView
         android:id="@+id/txt_no_comments"
