+++ app/src/main/res/navigation/nav_graph.xml
@@ -6,7 +6,7 @@
     app:startDestination="@id/albumFragment">
     <fragment
         android:id="@+id/albumFragment"
-        android:name="com.example.vinilosapp.ui.AlbumFragment"
+        android:name="com.example.vinilosapp.view.AlbumFragment"
         android:label="album_fragment"
         tools:layout="@layout/album_fragment" >
         <action
@@ -15,7 +15,7 @@
     </fragment>
     <fragment
         android:id="@+id/commentFragment"
-        android:name="com.example.vinilosapp.ui.CommentFragment"
+        android:name="com.example.vinilosapp.view.CommentFragment"
         android:label="comment_fragment"
         tools:layout="@layout/comment_fragment" >
         <argument
